Fix cast

+++ include/soul/patch/helper_classes/soul_patch_Utilities.h
@@ -188,7 +188,7 @@ inline choc::value::Value parseManifestFile (VirtualFile& manifestFile)
     }
     catch (choc::json::ParseError error)
     {
-        throwPatchLoadError (error.message, manifestPath, error.line, error.column);
+        throwPatchLoadError (error.message, manifestPath, static_cast<uint32_t> (error.line), static_cast<uint32_t> (error.column));
     }
 
     throwPatchLoadError ("Expected an object called '" + std::string (getManifestTopLevelPropertyName()) + "'", manifestPath);
